backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de données
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentification
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# E-mail
imapclient==3.0.1
python-dotenv==1.0.0
aiosmtplib==3.0.1

# Office 365 / Microsoft Graph
msal==1.25.0
httpx==0.25.2

# Gmail API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0

# IA
anthropic==0.7.7
openai==1.3.7

# Planification
apscheduler==3.10.4

# Utilitaires
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0
python-dateutil==2.8.2
pytz==2023.3

# Export
reportlab==4.0.7
openpyxl==3.1.2
pandas==2.1.3

# Logs
loguru==0.7.2

# Cache
redis==5.0.1

# Tests
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
